requirements.txt
crewai>=0.28.0
crewai-tools>=0.2.0
python-dotenv>=1.0.0
litellm>=1.30.0
ddgs>=0.1.0
requests>=2.31.0